requirements.txt
# Core LangChain (Aligned versions)
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0

# Vector Store & Embeddings
langchain-chroma==0.1.4
langchain-huggingface==0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Reranking
flashrank==0.2.4

# PDF Processing
pypdf>=3.0.0
langchain-text-splitters==0.3.0

# Tools
duckduckgo-search>=6.0.0

# CLI & Server
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.7.0
numpy>=1.26.0
requests>=2.31.0